Move radar processing into it's own folder directory, use the seperate process to interact with radar data. Add cmd line params for some functions like skipping video procesing or radar processing options. Add proper handling for the q to quite option
