+++ README.md
@@ -17,7 +17,7 @@ Currently in progress. But this project's goal is to create software that can be
 export PYTHONPATH="$PYTHONPATH:/workspaces/UAV-Object-Tracking/yolo_video_object_tracking"
 sudo apt-get install python3-tk
 
-python3 object_tracking.py --weights yolo_video_object_tracking/yolov7.pt --source yolo_video_object_tracking/video_samples/M0101.mp4 --conf-thres 0.4 --no-download --view-img
+python3 object_tracking.py --weights yolo_video_object_tracking/yolov7.pt --source data/video/M0101.mp4 --conf-thres 0.4 --no-download --view-img
 
 python object_tracking.py --radar-only
 python object_tracking.py --video-only
